Tidy ups (#28)
* Checks for permissions when creating/viewing/updating/deleting, remove privilege mostly implemented and tests mostly updated

* finally fixed the remove tests

* front end for remove

* added amazon stuff/updated nuget packages

* issue with requires authentication fixed, remove privilege fully implemented

* modal all working to confirm delete

* pr feedback

* removing code cov (hopefully)

* webpack npm updates

* giving up on npm madness

* does even webpack updates build

* builds failing so just reverting everything to get this through for now

+++ src/Service/Modules/GroupModule.cs
@@ -1,24 +1,29 @@
 ﻿namespace Linn.Authorisation.Service.Modules
 {
+    using System.Linq;
     using Common.Facade;
-    using Domain.Groups;
     using Facade;
-
-    using Linn.Authorisation.Domain.Services;
-
+    using Linn.Common.Authorisation;
+    using Linn.Production.Domain.LinnApps;
+    using Linn.Production.Service.Extensions;
     using Nancy;
     using Nancy.ModelBinding;
+    using Nancy.Security;
     using Resources;
 
     public sealed class GroupModule : NancyModule
     {
         private readonly IGroupService groupService;
+
         private readonly IPermissionService permissionService;
 
-        public GroupModule(IGroupService groupService, IPermissionService permissionService)
+        private readonly IAuthorisationService authorisationService;
+
+        public GroupModule(IGroupService groupService, IPermissionService permissionService, IAuthorisationService authorisationService)
         {
             this.groupService = groupService;
             this.permissionService = permissionService;
+            this.authorisationService = authorisationService;
             this.Post("/authorisation/groups", _ => this.CreateGroup());
             this.Get("/authorisation/groups", _ => this.GetGroups());
             this.Get("/authorisation/groups/{id:int}", parameters => this.GetGroup(parameters.id));
@@ -30,6 +35,14 @@
 
         private object CreateGroup()
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to create groups")));
+            }
+
             var resource = this.Bind<GroupResource>();
             var result = this.groupService.Add(resource);
             return this.Negotiate.WithModel(result);
@@ -37,12 +50,28 @@
 
         private object GetGroup(int id)
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to view groups")));
+            }
+
             var group = this.groupService.GetById(id);
             return this.Negotiate.WithModel(group);
         }
 
         private object UpdateGroup(int id)
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to edit groups")));
+            }
+
             var resource = this.Bind<GroupResource>();
             var result = this.groupService.Update(id, resource);
             return this.Negotiate.WithModel(result);
@@ -50,12 +79,28 @@
 
         private object GetGroups()
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to view groups")));
+            }
+
             var groups = this.groupService.GetAll();
             return this.Negotiate.WithModel(groups);
         }
-        
+
         private object AddGroupMember(int id)
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to edit groups")));
+            }
+
             var resource = this.Bind<GroupMemberResource>();
             var result = this.groupService.AddGroupMember(id, resource);
             return this.Negotiate.WithModel(result);
@@ -63,12 +108,28 @@
 
         private object RemoveGroupMember(int id, int memberId)
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to edit groups")));
+            }
+
             var result = this.groupService.RemoveGroupMember(id, memberId);
             return this.Negotiate.WithModel(result);
         }
 
         private object GetGroupPermissions(int id)
         {
+            this.RequiresAuthentication();
+            var privileges = this.Context?.CurrentUser?.GetPrivileges().ToList();
+
+            if (!this.authorisationService.HasPermissionFor(AuthorisedAction.AuthorisationAdmin, privileges))
+            {
+                return this.Negotiate.WithModel((new BadRequestResult<string>("You are not authorised to view groups")));
+            }
+
             var result = this.permissionService.GetImmediatePermissionsForGroup(id);
             return this.Negotiate.WithModel(result);
         }
